+++ frontend/src/components/ArenaView.jsx
@@ -3,32 +3,8 @@ import { socket, API_URL } from '../socket.js';
 import './ArenaView.css';
 import '../text_decor.css';
 
-export default function ArenaView() {
-  const defaultBattleState = {
-    fighters: [
-      {
-        name: "",
-        description: "",
-        wins: 0,
-        losses: 0,
-      },
-      {
-        name: "",
-        description: "",
-        wins: 0,
-        losses: 0,
-      }
-    ],
-  }
-  const [battleState, setBattleState] = useState(defaultBattleState);
-  const [lastWinner, setLastWinner] = useState("")
-  const [logState, setLogState] = useState([{description: "The match will begin soon!"}]);
-  const [summaryState, setSummaryState] = useState("");
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState(null);
-  const [timer, setTimer] = useState(null);
-  const timeouts = useRef([]);
-  
+export default function ArenaView({ battleState, timer, logState, lastWinner, summaryState}) {
+
   function ImageViewer({ base64, isWinner, isLoser }) {
     let className = '';
     if (isWinner) className = 'winner-img';
@@ -42,117 +18,6 @@ export default function ArenaView() {
     );
   }
 
-  function processFightData(data) {
-    // Processes fighting data
-    console.log(data);
-
-    // Fighter images
-    setBattleState(data);
-  }
-
-  function processFightLogs(data) {
-    // Processes only the logs
-    setLogState([]);
-    const LOG_DELAY = 1500 // 1.5 seconds between message
-    data.log.forEach((log, index) => {
-      const t = setTimeout(() => {
-        setLogState(prev => [...prev, log]);
-        const logContainer = document.querySelector('.logs');
-        if (logContainer) logContainer.scrollTop = logContainer.scrollHeight;
-      }, (index + 1) * LOG_DELAY);
-      timeouts.current.push(t);
-    });
-    //totalTime is how long to wait before the winner text and summary is displayed. Calculated by the amount of logs times the log delay
-    const totalTime = (data.log.length + 1) * LOG_DELAY;
-    const tWinner = setTimeout(() => {
-      setLastWinner(data.winner);
-      setSummaryState(data.summary);
-      setBattleState(data);
-    }, totalTime);
-    timeouts.current.push(tWinner);
-  }
-
-  const handleSchedule = (data) => {
-    console.log("SCHEDULE ------");
-    timeouts.current.forEach(clearTimeout);
-    timeouts.current = [];
-
-    setLogState([{description: "The match will begin soon!"}]);
-    setLastWinner("");
-    setSummaryState("");
-    processFightData(data);
-  }
-
-  const handleResult = (data) => {
-    // Takes data from a fight and places it in the correct places
-    console.log("RESULT ------")
-    //in order to preview the new character name and description we used a mixed battle state
-    //the mixed battle state retains all the information of the previous battle state but includes the new name and description
-    //this is only used temporarily, until the complete battle state will overwrite this inside of processFightLogs
-    setBattleState(prev => {
-      const mixedState = {...data, fighters: [...data.fighters]};
-      mixedState.fighters = data.fighters.map((newFighter, index) => {
-        const oldFighter = prev.fighters[index];
-        return {
-          ...newFighter,
-          wins: oldFighter ? oldFighter.wins : newFighter.wins,
-          losses: oldFighter ? oldFighter.losses : newFighter.losses
-        };
-      });
-      return mixedState;
-    });
-    processFightLogs(data);
-  }
-
-  const handleTimerUpdate = (data) => {
-    console.log("TIMER UPDATE ------")
-    if (data.time_left > 0) {
-      setTimer(data.time_left);
-    } 
-    else if (data.time_left == 0) {
-      setTimer("Battle commencing!")
-    }
-    else {
-      setTimer("Bookie is working on the next match...")
-    }
-    console.log(data);
-  }
-
-  // Listen for and load battles from backend
-  useEffect(() => {
-    // Register listeners
-    socket.on('match_scheduled', handleSchedule);
-    socket.on('match_result', handleResult);
-    socket.on('timer_update', handleTimerUpdate);
-
-    // Get initial fighter info from scheduled battle
-    fetch(`${API_URL}/api/card`)
-      .then(response => {
-        if (!response.ok) throw new Error('Network response was not ok');
-        return response.json();
-      })
-      .then(json => {
-        console.log("Got Fresh Fighter Data");
-        processFightData(json);
-        setLoading(false);
-      })
-      .catch(err => {
-        setError(err.message);
-        setLoading(false);
-      });
-
-
-    // De-register listeners for cleanup
-    return () => {
-      socket.off('match_scheduled', handleSchedule);
-      socket.off('match_result', handleResult);
-      socket.off('timer_update', handleTimerUpdate);
-    }
-  }, []);
-
-
-  if (loading) return <div class='net-loading'>Loading...</div>;
-  if (error) return <div class='net-error'>Error: {error}</div>;
   if (!battleState || !battleState.fighters || battleState.fighters.length < 2) { return (
         <div className='root waiting-screen'>
             <h1>Waiting for Next Match...</h1>
